+++ app/src/main/java/com/example/trainticketproject/daos/UserDAO.java
@@ -18,13 +18,13 @@ public interface UserDAO {
     void insert(User user);
 
     @Query("SELECT * FROM User WHERE uid = :uid")
-    LiveData<User> getUserById(Integer uid);
+    LiveData<User> getUserById(Long uid);
 
     @Insert
     void insertMultipleUser(List<User> users);
 
     @Query("UPDATE User SET rewardPoint = :newRewardPoint WHERE uid = :uid")
-    void updateRewardPoint(int newRewardPoint, Integer uid);
+    void updateRewardPoint(int newRewardPoint, Long uid);
 
     @Query("SELECT * FROM User")
     LiveData<List<User>> getAllUsers();
